sort out files and components
