+++ src/components/icon/addIcon.png
