+++ src/components/icon/okIcon.png
