+++ src/components/icon/trashCan.png
