+++ src/components/icon/trashCanHover.png
